notification_client/requirements.txt
python-socketio[client]>=5.11.0
websocket-client>=1.6.0
win10toast-click>=0.1.2
requests>=2.28.0

